complete new function in user respository

+++ server/model/user_type.go
@@ -63,7 +63,27 @@ type OrderOperation struct {
 }
 
 type BalanceHistory struct {
-	Timestamp uint    `bson:"timestamp" json:"timestamp"`
+	Timestamp int64    `bson:"timestamp" json:"timestamp"`
 	Balance   float64 `bson:"balance" json:"balance"`
 	Method    string  `bson:"method" json:"method"`
 }
+
+	// var projection bson.M
+	// if method == "DEPOSIT" || method == "WITHDRAW" {
+	// 	projection = bson.M{
+	// 		"balanceHistory": bson.M{
+	// 			"$elemMatch": bson.M{
+	// 				"method": "DEPOSIT",
+	// 			},
+	// 			"$slice": []int{0, 10},
+	// 		},
+	// 	}
+	// } else if method == "ALL" {
+	// 	projection = bson.M{
+	// 		"balanceHistory": bson.M{
+	// 			"$slice": []int{0, 10},
+	// 		},
+	// 	}
+	// } else {
+	// 	return []BalanceHistory{}, errors.New("invalid method")
+	// }
